+++ src/test/java/sns/pinocchio/domain/auth/AuthServiceTest.java
@@ -1,5 +1,9 @@
 package sns.pinocchio.domain.auth;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,19 +16,10 @@ import org.springframework.transaction.annotation.Transactional;
 import sns.pinocchio.PinocchioApplication;
 import sns.pinocchio.application.auth.AuthService;
 import sns.pinocchio.application.member.MemberService;
-import sns.pinocchio.application.member.memberDto.request.SignupRequestDto;
-import sns.pinocchio.application.member.memberDto.request.UpdateRequestDto;
-import sns.pinocchio.application.report.ReportService;
 import sns.pinocchio.config.global.auth.util.TokenProvider;
 import sns.pinocchio.domain.member.Member;
-import sns.pinocchio.domain.report.Report;
 import sns.pinocchio.infrastructure.member.MemberRepository;
 import sns.pinocchio.presentation.member.exception.MemberException;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static sns.pinocchio.domain.report.ReportedType.POST;
 
 @SpringBootTest(classes = PinocchioApplication.class)
 @AutoConfigureMockMvc
@@ -40,8 +35,6 @@ public class AuthServiceTest {
 
   @Autowired private PasswordEncoder passwordEncoder;
 
-  @Autowired private ReportService reportService;
-
   @Autowired private TokenProvider tokenProvider;
 
   @BeforeEach
@@ -55,35 +48,6 @@ public class AuthServiceTest {
             .build();
 
     memberRepository.save(member);
-  }
-
-  @Test
-  @DisplayName("계정 생성 테스트")
-  public void createMemberTest() {
-    SignupRequestDto member =
-        new SignupRequestDto("name", "example1@naver.com", "nickname1", "memberPassword123!@");
-    memberService.createMember(member);
-
-    Member user = memberService.findByEmail("example1@naver.com");
-
-    assertThat(user).isNotNull();
-    assertThat(user.getName()).isEqualTo("name");
-    assertThat(user.getNickname()).isEqualTo("nickname1");
-    assertThat(user.getTsid()).isNotNull();
-  }
-
-  @Test
-  @DisplayName("이메일 검증 성공 테스트")
-  public void validateEmail() {
-    assertThatCode(() -> memberService.findByEmail("example@naver.com")).doesNotThrowAnyException();
-  }
-
-  @Test
-  @DisplayName("존재하지 않는 이메일로 검증 실패 테스트")
-  public void validateEmail_Failure() {
-    assertThatThrownBy(() -> memberService.findByEmail("notfound@naver.com"))
-        .isInstanceOf(MemberException.class)
-        .hasMessageContaining("존재하지 않는 사용자입니다.");
   }
 
   @Test
@@ -106,50 +70,14 @@ public class AuthServiceTest {
   }
 
   @Test
-  @DisplayName("패스워드 변경 테스트")
-  public void modifyPasswordTest() {
-    Member member = memberService.findByEmail("example@naver.com");
-
-    memberService.changePassword(member, "abcd123@");
-
-    assertThatCode(() -> authService.validatePassword("abcd123@", member))
-        .doesNotThrowAnyException();
-  }
-
-  @Test
-  @DisplayName("프로필 수정 성고 테스트")
-  public void modifyProfile() {
-    Member member = memberService.findByEmail("example@naver.com");
-
-    UpdateRequestDto updateRequestDto =
-        new UpdateRequestDto("bob", "bob", "me", "null", "null", true);
-    member.updateProfile(updateRequestDto);
-
-    Member updateMember = memberService.findByNickname("bob");
-
-    assertThat(updateMember.getName()).isEqualTo("bob");
-    assertThat(updateMember.getBio()).isEqualTo("me");
-    assertThat(updateMember.getIsActive()).isTrue();
-  }
-
-  @Test
-  @DisplayName("신고 내역 저장 테스트")
-  public void createReportTest() {
-    reportService.createReport(1L, 2L, POST, "욕햇어용");
-    Report report = reportService.findByReporter(1L);
-
-    assertThat(report.getReportedId()).isEqualTo(2L);
-    assertThat(report.getReportedType()).isEqualTo(POST);
-    assertThat(report.getReason()).isEqualTo("욕햇어용");
-  }
-
-  @Test
   @DisplayName("토큰 생성 테스트")
   public void createToken() {
     Member member = memberService.findByEmail("example@naver.com");
 
     String accessToken = tokenProvider.generateAccessToken(member);
+    String refreshToken = tokenProvider.generateRefreshToken();
 
     assertThat(accessToken).isNotNull();
+    assertThat(refreshToken).isNotNull();
   }
 }
