+++ src/test/java/sns/pinocchio/domain/member/MemberServiceTest.java
@@ -9,11 +9,19 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.transaction.annotation.Transactional;
 import sns.pinocchio.PinocchioApplication;
+import sns.pinocchio.application.auth.AuthService;
 import sns.pinocchio.application.member.MemberService;
+import sns.pinocchio.application.member.memberDto.request.SignupRequestDto;
 import sns.pinocchio.application.member.memberDto.request.UpdateRequestDto;
+import sns.pinocchio.application.report.ReportService;
+import sns.pinocchio.domain.report.Report;
 import sns.pinocchio.infrastructure.member.MemberRepository;
+import sns.pinocchio.presentation.member.exception.MemberException;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static sns.pinocchio.domain.report.ReportedType.POST;
 
 @SpringBootTest(classes = PinocchioApplication.class)
 @AutoConfigureMockMvc
@@ -22,8 +30,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class MemberServiceTest {
 
   @Autowired private MemberService memberService;
-
   @Autowired private MemberRepository memberRepository;
+  @Autowired private ReportService reportService;
+  @Autowired private AuthService authService;
 
   @BeforeEach
   public void init() {
@@ -39,14 +48,32 @@ public class MemberServiceTest {
   }
 
   @Test
-  @DisplayName("유저 조회 테스트")
-  public void getMemberTest() {
-    Member member = memberService.findByEmail("example1@naver.com");
+  @DisplayName("계정 생성 테스트")
+  public void createMemberTest() {
+    SignupRequestDto member =
+        new SignupRequestDto("name", "example1@naver.com", "nickname1", "memberPassword123!@");
+    memberService.createMember(member);
 
-    assertThat(member.getName()).isEqualTo("member");
-    assertThat(member.getEmail()).isEqualTo("example1@naver.com");
-    assertThat(member.getNickname()).isEqualTo("nickname1");
-    assertThat(member.getIsActive()).isFalse();
+    Member user = memberService.findByEmail("example1@naver.com");
+
+    assertThat(user).isNotNull();
+    assertThat(user.getName()).isEqualTo("name");
+    assertThat(user.getNickname()).isEqualTo("nickname1");
+    assertThat(user.getTsid()).isNotNull();
+  }
+
+  @Test
+  @DisplayName("이메일 검증 성공 테스트")
+  public void validateEmail() {
+    assertThatCode(() -> memberService.findByEmail("example@naver.com")).doesNotThrowAnyException();
+  }
+
+  @Test
+  @DisplayName("존재하지 않는 이메일로 검증 실패 테스트")
+  public void validateEmail_Failure() {
+    assertThatThrownBy(() -> memberService.findByEmail("notfound@naver.com"))
+        .isInstanceOf(MemberException.class)
+        .hasMessageContaining("존재하지 않는 사용자입니다.");
   }
 
   @Test
@@ -65,5 +92,27 @@ public class MemberServiceTest {
     assertThat(result.getWebsite()).isEqualTo("youtube");
     assertThat(result.getProfileImageUrl()).isEmpty();
     assertThat(result.getIsActive()).isFalse();
+  }
+
+  @Test
+  @DisplayName("패스워드 변경 테스트")
+  public void modifyPasswordTest() {
+    Member member = memberService.findByEmail("example@naver.com");
+
+    memberService.changePassword(member, "abcd123@");
+
+    assertThatCode(() -> authService.validatePassword("abcd123@", member))
+        .doesNotThrowAnyException();
+  }
+
+  @Test
+  @DisplayName("신고 내역 저장 테스트")
+  public void createReportTest() {
+    reportService.createReport(1L, 2L, POST, "욕햇어용");
+    Report report = reportService.findByReporter(1L);
+
+    assertThat(report.getReportedId()).isEqualTo(2L);
+    assertThat(report.getReportedType()).isEqualTo(POST);
+    assertThat(report.getReason()).isEqualTo("욕햇어용");
   }
 }
